refactor: extract logger context functions and replace appendRunnerName with WithRunnerDepth
